Accurate percentage added. All complications added.
The accurate percentage will now be sent together with each complications, in addition the severity_string is updated with this accurate percentage. Also the newest complications is added and will be sent to frontend. (They are still missing most of their factors).

The old percantage_risk might not be usefull anymore. This field should probably be removed from the database.